Set the compiler target to 1.8
Although there is no need for a compiler for this project, this will
ensure that Eclipse will not display a message stating that there is no
strictly compatible version of the JDK assuming JDK8 is installed.

With JDK7 updates from Oracle already discontinued, JDK8 will be the
target for non-compiled artifacts going forward.